Fix trigger report when query fails
In the old implementation, when dropping the database, all the queries
failed with trigger report because the tables had not been yet created.

Use trigger_report() only after UpdateAll is done, to be sure that the
tables are created at that time.

Also don't trigger report on warning if the query is 'drop table ...'
because the database will not be created and other fatal errors will be
triggered because of that.