Improve display icons for the hidden note case
If the $show_notes=false then the note is always hidden.
To make a difference when user adds a new note the comment icon is
changes and looks as it is full, but when a field has no note then
the comment icon looks empty.
Also fix the id of the advanced field that is show/hidden.